Enable use of boto native access resolution and make role assumption optional
Previously the default region to look for the CloudTrail SQS queue in was set
to us-west-1 as a default in the es_worker. This is now set to '' in the
es_worker and as a result will default to whatever region boto determines

This commit introduces a new function, get_aws_credentials, which accepts AWS
API key arguments, checks if they are set to defaults (either defaults in the
code or in the conf file) and if they are, does not return them. This enables
you to conditionally pass or not pass API keys to boto functions depending on
whether or not they were set to something other than the defaults. The result
is that by not setting API keys, MozDef will instead rely on the boto access
resoltion methods which include checking local ~/.aws/ files as well as
instance metadata.

This commit also allows the `cloudtrail_arn` variable, which is actually the
ARN of an IAM role, to be optional. If the value is set to the default, it is
ignored and no IAM Role Assumption is done when attempting to fetch data from
the S3 bucket. Instead the native credentials are used.