Only allow one query to be configured.
When there are multiple queries, GeoModel will try to avoid processing
events that it may have already processed in previous iterations.
Rather than building a tagging system to track this internally, we will
restrict the number of queries that can be run so that this condition is
impossible.  Queries against multiple sources can be written using OR,
such as in "tags:auth0 OR tags:duo"