Disable use of codeowners in GitHub
As /cloudy_mozdef/ isn't being worked on anymore, there's little need to enable
Gene specific ability to review code in that directory structure. The way GitHub
implements CODEOWNERS, the result is that Gene gets asked for review on any PR
which contains a change to CHANGELOG.md or /docs/ which is most PRs, however his
review is not sufficient for merge as the PR would contain other changes as well.

By removing CODEOWNERS the repo goes back to expecing review from those authorized
and not differentiating areas in the codebase